+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [2-alpha8]  - 2022-11-10
+- adapting to Kindly changes - returning multiple contexts
+
 ## [2-alpha8]  - 2022-11-06
 - updated Kindly deps